Add isEventSupported(type) + more
The Y.Event default event now supports subscribing to and detaching from
custom events if the subscribed event name isn't in Y.Node.NODE_EVENTS or,
if 'node' isn't loaded, Y.Event.isEventSupported(type) returns false.

It will be a rare case that 'node' won't be loaded so this is a bit of code
weight, but the complexity should be about the same.

Also cleaned up the subscribe and unsubscribe code and made unsubscribe remove
the DOM event listener when the last subscriber is removed.