Move default logic from Y.Event to DOMEvent
The default logic that was added to Y.Event's default event was moved to
DOMEvent, the CustomEvent that handles the events in the DOM_EVENT whitelist.
As a result, it could be trimmed down considerably.

Much cleaner implementation, though I'm starting to get concerned about the
number of function hops before substantive code executes.