Add eventx-onload for Y.on('load', winLoad)
Published as a fireOnce event on Y.Global, then the DOM event is overridden to
subscribe to the custom event instead for

* Y.on('load', callback);
* Y.on('load', callback, Y.config.win);
* Y.one('win').on('load', callback);
* Y.one(Y.config.win).on('load', callback);

Callback context override currently defaults to Y.Global. Not sure what it
should default to. Y? Extra args are supported.

The custom event currently fires without passing the DOM event as the callback
arg. If it's needed, I can add it.