Remove most of synthetic logic, now unnecessary
Logic absorbed into FacadeEvent or DOMEvent. Very little shimming was necessary
for backward compatibility.

There is one critical incompatibility: if a synth implementation defines an
on() or delegate() that calls this.on(), it will result in an infinite loop,
because on() is supplied by SyntheticEvent, and configured on() and delegate()
are moved to _oldOn() and _oldDelegate(), which are called from the core on().